Issue #4 - Added upgrade command, error handlers
Built-out the upgrade command, which ends up being a simple variation on "install". The main difference is that if no packages are specified we upgrade everything in the local index.

Also added error handlers for install/reinstall. Will experiment with this a bit, if there is an error do we truly want to try to keep going?