Keep local versions later than local versions
If a local version is more recent than a remote version keep the local version of an installed package. This seems like the most correct behavior, instead of just detecting if the version is equal to remote. A bit of a pathological case but it could be useful when doing bleeding edge development work.